+++ src/main/java/com/springAdvanced/Telephonedirectory/repository/CompanyRepository.java
@@ -3,6 +3,6 @@ package com.springAdvanced.Telephonedirectory.repository;
 import com.springAdvanced.Telephonedirectory.model.Company;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CompanyRepository extends CrudRepository<Company, Integer>
+public interface CompanyRepository extends CrudRepository<Company, Long>
 {
 }
